Create subcommand apply
